% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-extraction.R
\name{extractTables}
\alias{extractTables}
\title{Extract tables from web pages.}
\usage{
extractTables(url, findHeaders = TRUE, mergeTables = TRUE,
  tableXPath = "//table", rowTag = "tr", cellTag = c("th", "td"))
}
\arguments{
\item{url}{A web page URL.}

\item{findHeaders}{Whether to guess table headers based on XML tags and
attributes.}

\item{mergeTables}{Whether to merge tables with the same schema.}

\item{tableXPath}{XPath expression for extracting table nodes.}

\item{rowTag}{Table row XML tag(s).}

\item{cellTag}{Table cell XML tag(s).}
}
\value{
A list of tables extracted from the web page.
}
\description{
The function allows to extract tabular data from a web page. The extraction
function can perform additional cleaning and inference:
\itemize{
  \item taking care of "span" attributes
  \item guessing the table headers based on tags' names and attributes
  \item merging tables with the same schema
}
}
\examples{
aircrafts <- extractTables("http://www.flugzeuginfo.net/table_accodes_en.php")
cssProperties <- extractTables("http://www.w3schools.com/cssref/")
}

