% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-extraction.R
\name{findHeaderCount}
\alias{findHeaderCount}
\title{Guess header count}
\usage{
findHeaderCount(table, rowTag = "tr", cellTag = c("th", "td"))
}
\arguments{
\item{table}{A table XML node.}

\item{rowTag}{Table row XML tag(s).}

\item{cellTag}{Table cell XML tag(s).}
}
\value{
The number of the first rows that represent the \code{table} header.
}
\description{
The function uses XML tags to guess how many table rows form the table headers.
}
\examples{
findHeaderCount(table)
}

