% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-merger.R
\name{mergeTables}
\alias{mergeTables}
\title{Merge data frames}
\usage{
mergeTables(tables)
}
\arguments{
\item{tables}{A list of data frames.}
}
\value{
A list of data frames, each of which is one of the original data
frames from \code{tables} or a union of the original data frames.
}
\description{
The function merges data frames based on column names.
}
\examples{
mergeTables(tables)
}

