% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-extraction.R
\name{setHeaders}
\alias{setHeaders}
\title{Set column names}
\usage{
setHeaders(data, headerCount = 1)
}
\arguments{
\item{data}{A data frame.}

\item{headerCount}{How many rows use as headers.}
}
\value{
A data frame with the first \code{headerCount} rows turned into
column names.
}
\description{
The function sets column names in a data frame using data from the first rows.
}
\examples{
df <- data.frame(letters[1:5], as.character(seq(7, 11)), letters[13:17])
dfWithColnames <- setHeaders(df)
dfWithColnames <- setHeaders(df, headerCount = 2)
}

