% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-tools.R
\name{tableToDataFrame}
\alias{tableToDataFrame}
\title{Convert XML node to data frame}
\usage{
tableToDataFrame(table, findHeaders = TRUE, rowTag = "tr",
  cellTag = c("th", "td"))
}
\arguments{
\item{table}{XML node.}

\item{findHeaders}{Whether to guess table headers based on XML tags and
attributes.}

\item{rowTag}{Table row XML tag(s).}

\item{cellTag}{Table cell XML tag(s).}
}
\value{
A data frame containing the data from the original XML node.
}
\description{
The function converts "table" XML node to a data frame.
}

